add tests remove types file
