+++ assignments/helpers/helpers.go
@@ -3,5 +3,5 @@ package helpers
 import "strings"
 
 func RemoveNewlineFromStr(str string) string {
-	return strings.Replace(str, "\n", "", 1)
+	return strings.Replace(str, "\n", "", -1)
 }
